(1584) mutualisation serializeLivingConditions into frontend/common
