+++ packages/frontend/common/helpers/town/living_conditions/toilet_types.js
